+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/EditorScripts/SchemaConfigurationAdditionalDocumentationPROEditorScript.php
@@ -1,0 +1,68 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GraphQLAPI\GraphQLAPI\Services\EditorScripts;
+
+use GraphQLAPI\GraphQLAPI\ModuleResolvers\UserInterfaceFunctionalityModuleResolver;
+use GraphQLAPI\GraphQLAPI\Services\CustomPostTypes\GraphQLSchemaConfigurationCustomPostType;
+use GraphQLAPI\GraphQLAPI\Services\Scripts\MainPluginScriptTrait;
+
+class SchemaConfigurationAdditionalDocumentationPROEditorScript extends AbstractEditorScript
+{
+    use MainPluginScriptTrait;
+
+    private ?GraphQLSchemaConfigurationCustomPostType $graphQLSchemaConfigurationCustomPostType = null;
+
+    final public function setGraphQLSchemaConfigurationCustomPostType(GraphQLSchemaConfigurationCustomPostType $graphQLSchemaConfigurationCustomPostType): void
+    {
+        $this->graphQLSchemaConfigurationCustomPostType = $graphQLSchemaConfigurationCustomPostType;
+    }
+    final protected function getGraphQLSchemaConfigurationCustomPostType(): GraphQLSchemaConfigurationCustomPostType
+    {
+        /** @var GraphQLSchemaConfigurationCustomPostType */
+        return $this->graphQLSchemaConfigurationCustomPostType ??= $this->instanceManager->getInstance(GraphQLSchemaConfigurationCustomPostType::class);
+    }
+
+    protected function getScriptName(): string
+    {
+        return 'schema-configuration-additional-documentation-pro';
+    }
+
+    public function getEnablingModule(): ?string
+    {
+        return UserInterfaceFunctionalityModuleResolver::SCHEMA_CONFIGURATION_ADDITIONAL_DOCUMENTATION_PRO;
+    }
+
+    /**
+     * Add the locale language to the localized data?
+     */
+    protected function addLocalLanguage(): bool
+    {
+        return true;
+    }
+
+    /**
+     * Default language for the script/component's documentation
+     */
+    protected function getDefaultLanguage(): ?string
+    {
+        // English
+        return 'en';
+    }
+
+    protected function getAllowedPostTypes(): array
+    {
+        return array_merge(
+            parent::getAllowedPostTypes(),
+            [
+                $this->getGraphQLSchemaConfigurationCustomPostType()->getCustomPostType(),
+            ]
+        );
+    }
+
+    protected function registerStyleIndexCSS(): bool
+    {
+        return true;
+    }
+}